Add basic query parsing and serialisation
This adds a bulk of the "core" supporting code to parse and serialize
binary DNS packets, converting various bits of data from them into rich
structs that we can operate on, and then turn back into binary packets
when we're done.

This is currently working with dig(1), albeit only with the one stub
hostname that's currently hardcoded:

        % dig +short @127.0.0.1 -p 5010 +noedns www.example.com
        127.0.0.1

Right now EDNS is not supported at all, though we just ignore it, so
technically it "works", but because we don't reserialize it DNS clients
will be angry. EDNS is a requirement for MVP.

Speaking of, next up:

- EDNS support
- AAAA support
- Error message responses
- Actual zone database
- Lots of other stuff far off (dnssec? DoH? DoTLS? axfr?) maybe