Implement naive name compression
This is a very simple implementation of the name compression scheme as
outlined by the RFC. Instead of a name, we provide two `1` bits followed
by a pointer to the number of bytes from the beginning of the packet
where we can find the name.

In this case, we're not doing any searching, we're just assuming we can
find it after the first 12 bytes of the header. If it's not there we
fall back to not using compression. This might not be the most
efficient, technically, as the string may appear elsewhere in the
packet, but it will work. We can look at making it "smarter" later.