Use erlang inet-compatible type and class atoms
Instead of the human strings, use atoms that follow the same format as
Erlang's `inet_res` module types. This will make interfacing with things
easier in the future, especially if we need to fall on some inet
functionality. Plus this feels more "correct" for the struct itself --
human string mappings are a display concern, not a data one.